% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_create_application}
\alias{codedeploy_create_application}
\title{Creates an application}
\usage{
codedeploy_create_application(applicationName, computePlatform)
}
\arguments{
\item{applicationName}{[required] The name of the application. This name must be unique with the
applicable IAM user or AWS account.}

\item{computePlatform}{The destination platform type for the deployment (\code{Lambda} or \code{Server}).}
}
\description{
Creates an application.
}
\section{Request syntax}{
\preformatted{svc$create_application(
  applicationName = "string",
  computePlatform = "Server"|"Lambda"|"ECS"
)
}
}

\keyword{internal}
