% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedeploy_operations.R
\name{codedeploy_get_deployment_group}
\alias{codedeploy_get_deployment_group}
\title{Gets information about a deployment group}
\usage{
codedeploy_get_deployment_group(applicationName, deploymentGroupName)
}
\arguments{
\item{applicationName}{[required] The name of an AWS CodeDeploy application associated with the IAM user
or AWS account.}

\item{deploymentGroupName}{[required] The name of a deployment group for the specified application.}
}
\description{
Gets information about a deployment group.
}
\section{Request syntax}{
\preformatted{codedeploy$get_deployment_group(
  applicationName = "string",
  deploymentGroupName = "string"
)
}
}

\keyword{internal}
