% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_get_job_details}
\alias{codepipeline_get_job_details}
\title{Returns information about a job}
\usage{
codepipeline_get_job_details(jobId)
}
\arguments{
\item{jobId}{[required] The unique system-generated ID for the job.}
}
\description{
Returns information about a job. Only used for custom actions.
}
\details{
When this API is called, AWS CodePipeline returns temporary credentials
for the Amazon S3 bucket used to store artifacts for the pipeline, if
the action requires access to that Amazon S3 bucket for input or output
artifacts. Additionally, this API returns any secret values defined for
the action.
}
\section{Request syntax}{
\preformatted{svc$get_job_details(
  jobId = "string"
)
}
}

\keyword{internal}
