% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_get_pipeline_state}
\alias{codepipeline_get_pipeline_state}
\title{Returns information about the state of a pipeline, including the stages
and actions}
\usage{
codepipeline_get_pipeline_state(name)
}
\arguments{
\item{name}{[required] The name of the pipeline about which you want to get information.}
}
\description{
Returns information about the state of a pipeline, including the stages
and actions.
}
\details{
Values returned in the revisionId and revisionUrl fields indicate the
source revision information, such as the commit ID, for the current
state.
}
\section{Request syntax}{
\preformatted{svc$get_pipeline_state(
  name = "string"
)
}
}

\keyword{internal}
