% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepipeline_operations.R
\name{codepipeline_put_job_success_result}
\alias{codepipeline_put_job_success_result}
\title{Represents the success of a job as returned to the pipeline by a job
worker}
\usage{
codepipeline_put_job_success_result(jobId, currentRevision,
  continuationToken, executionDetails)
}
\arguments{
\item{jobId}{[required] The unique system-generated ID of the job that succeeded. This is the
same ID returned from PollForJobs.}

\item{currentRevision}{The ID of the current revision of the artifact successfully worked upon
by the job.}

\item{continuationToken}{A token generated by a job worker, such as an AWS CodeDeploy deployment
ID, that a successful job provides to identify a custom action in
progress. Future jobs will use this token in order to identify the
running instance of the action. It can be reused to return additional
information about the progress of the custom action. When the action is
complete, no continuation token should be supplied.}

\item{executionDetails}{The execution details of the successful job, such as the actions taken
by the job worker.}
}
\description{
Represents the success of a job as returned to the pipeline by a job
worker. Only used for custom actions.
}
\section{Request syntax}{
\preformatted{svc$put_job_success_result(
  jobId = "string",
  currentRevision = list(
    revision = "string",
    changeIdentifier = "string",
    created = as.POSIXct(
      "2015-01-01"
    ),
    revisionSummary = "string"
  ),
  continuationToken = "string",
  executionDetails = list(
    summary = "string",
    externalExecutionId = "string",
    percentComplete = 123
  )
)
}
}

\keyword{internal}
