% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_groups}
\alias{xray_get_groups}
\title{Retrieves all active group details}
\usage{
xray_get_groups(NextToken)
}
\arguments{
\item{NextToken}{Pagination token. Not used.}
}
\description{
Retrieves all active group details.
}
\section{Request syntax}{
\preformatted{svc$get_groups(
  NextToken = "string"
)
}
}

\keyword{internal}
