% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_sampling_rules}
\alias{xray_get_sampling_rules}
\title{Retrieves all sampling rules}
\usage{
xray_get_sampling_rules(NextToken)
}
\arguments{
\item{NextToken}{Pagination token. Not used.}
}
\description{
Retrieves all sampling rules.
}
\section{Request syntax}{
\preformatted{svc$get_sampling_rules(
  NextToken = "string"
)
}
}

\keyword{internal}
