% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_trace_graph}
\alias{xray_get_trace_graph}
\title{Retrieves a service graph for one or more specific trace IDs}
\usage{
xray_get_trace_graph(TraceIds, NextToken)
}
\arguments{
\item{TraceIds}{[required] Trace IDs of requests for which to generate a service graph.}

\item{NextToken}{Pagination token. Not used.}
}
\description{
Retrieves a service graph for one or more specific trace IDs.
}
\section{Request syntax}{
\preformatted{svc$get_trace_graph(
  TraceIds = list(
    "string"
  ),
  NextToken = "string"
)
}
}

\keyword{internal}
