% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_get_trace_summaries}
\alias{xray_get_trace_summaries}
\title{Retrieves IDs and metadata for traces available for a specified time
frame using an optional filter}
\usage{
xray_get_trace_summaries(StartTime, EndTime, Sampling, FilterExpression,
  NextToken)
}
\arguments{
\item{StartTime}{[required] The start of the time frame for which to retrieve traces.}

\item{EndTime}{[required] The end of the time frame for which to retrieve traces.}

\item{Sampling}{Set to \code{true} to get summaries for only a subset of available traces.}

\item{FilterExpression}{Specify a filter expression to retrieve trace summaries for services or
requests that meet certain requirements.}

\item{NextToken}{Specify the pagination token returned by a previous request to retrieve
the next page of results.}
}
\description{
Retrieves IDs and metadata for traces available for a specified time
frame using an optional filter. To get the full traces, pass the trace
IDs to \code{BatchGetTraces}.
}
\details{
A filter expression can target traced requests that hit specific service
nodes or edges, have errors, or come from a known user. For example, the
following filter expression targets traces that pass through
\code{api.example.com}:

\code{service("api.example.com")}

This filter expression finds traces that have an annotation named
\code{account} with the value \code{12345}:

\code{annotation.account = "12345"}

For a full list of indexed fields and keywords that you can use in
filter expressions, see \href{http://docs.aws.amazon.com/xray/latest/devguide/xray-console-filters.html}{Using Filter Expressions}
in the \emph{AWS X-Ray Developer Guide}.
}
\section{Request syntax}{
\preformatted{svc$get_trace_summaries(
  StartTime = as.POSIXct(
    "2015-01-01"
  ),
  EndTime = as.POSIXct(
    "2015-01-01"
  ),
  Sampling = TRUE|FALSE,
  FilterExpression = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
