% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_operations.R
\name{xray_update_group}
\alias{xray_update_group}
\title{Updates a group resource}
\usage{
xray_update_group(GroupName, GroupARN, FilterExpression)
}
\arguments{
\item{GroupName}{The case-sensitive name of the group.}

\item{GroupARN}{The ARN that was generated upon creation.}

\item{FilterExpression}{The updated filter expression defining criteria by which to group
traces.}
}
\description{
Updates a group resource.
}
\section{Request syntax}{
\preformatted{svc$update_group(
  GroupName = "string",
  GroupARN = "string",
  FilterExpression = "string"
)
}
}

\keyword{internal}
